requirements.txt
# Project dependencies
ruff>=0.14.0
pytest>=8.4.0
pre-commit>=4.3.0
python-dotenv>=1.2.0
scipy>=1.11.0
dspy-ai>=2.0.0
